Updating to reflect HTTPS requirement.
These instructions were previously incomplete. Users need to enable HTTPS in Vorlon to use it with add-ins. I updated the instructions to include full details for how to do this. The instructions for enabling the Office plugin for Vorlon are greatly simplified now as well, since the add-in is included with the master branch of Vorlon (so users just need to change the enabled property to true).