Update troubleshoot-manifest.md
Updated this to include information on the Office Add-in Validator. Also, removed Visual Studio from the schema validation section, because this is no longer accurate now that VS pulls from the same validation service as the Office Add-in Validator.

Wasn't 100% sure how to link from the list of methods at the top to the specific section, so I need a bit of assistance there.