Update office.js CDN URL to all lowercase letters
The CDN URL for office.js should be changed to all lowercase letters, because many of our tools and samples are referencing it that way. Furthermore, we are now prefetching the office.js using the lowercase URL in Office Online. Using uppercase O in the URL causes us to be unable to share the browser cache as the URLs are different.