[Excel] (Custom functions) Add http to 'helpUrl' (#3880)
The code example provided for the @helpurl parameter does not use the "http://" prefix. If the developer implements it like this it won't work.

The documentation for manually creating the functions.json file correctly indicates the use of the "http://" prefix. See https://learn.microsoft.com/en-us/office/dev/add-ins/excel/custom-functions-json#json-metadata-example